Unit tests for use cases

+++ app/src/main/java/com/github/fatihsokmen/savingsgoals/data/model/SavingsGoalDto.kt
@@ -12,5 +12,5 @@ data class SavingsGoalDto(
 @Serializable
 data class TotalSavedDto(
     val currency: String,
-    val minorUnits: Long
+    val minorUnits: Double
 )
